fix(spark): fix launch reverts, close pool-squat DoS, native-only fees
Root cause of live launch reverts (two independent bugs):
- One-sided tick range branches in _mintAndRegister were inverted
  relative to V3 liquidity math, making mint() always compute zero
  liquidity and revert.
- slot0()'s feeProtocol was declared uint8 (vanilla Uniswap V3) but
  PancakeSwap V3 packs it as uint32; ABI decoding reverted whenever a
  pool's real value exceeded 255.

Security hardening:
- Pool-address squatting DoS: _clone() used CREATE off the launcher's
  nonce, letting anyone predict and pre-squat future pool addresses for
  free via the permissionless factory. Switched to CREATE2 salted by
  (sender, timestamp, launch params), and adopt-if-uninitialized instead
  of reverting when a pool shell already exists.
- _safeApprove now checks the decoded return value, not just call
  success, matching the rest of the file's ERC-20 handling.
- Added InvalidTickRange guard so a misconfigured marketCapRef fails
  with a clear error instead of an opaque downstream DEX revert.

Tokenomics: seed 100% of supply one-sided (was 99.99% with a 0.01%
creator reserve); no deliberate allocation, only mint-rounding dust is
swept to the creator.

Fee model: launch fee is now a single global value charged in native
ETH/BNB on every launch, regardless of quote token. The quote token is
now purely the pool's pairing partner — no ERC-20 is ever pulled from
the creator. Any ETH above the fee funds an instant buy: direct swap
when the quote token is WETH, multihop (WETH -> quoteToken -> spark)
through existing liquidity otherwise. Removed the now-dead decimals
field and _pullFrom helper.

SPARK.md updated throughout to match.

+++ spark/SparkLocker.sol
@@ -31,7 +31,7 @@ interface IUniswapV3PoolFees {
     function slot0() external view returns (
         uint160 sqrtPriceX96, int24 tick, uint16 observationIndex,
         uint16  observationCardinality, uint16 observationCardinalityNext,
-        uint8   feeProtocol, bool unlocked
+        uint32  feeProtocol, bool unlocked // PancakeSwap V3 packs this wider than uint8
     );
     function feeGrowthGlobal0X128() external view returns (uint256);
     function feeGrowthGlobal1X128() external view returns (uint256);
